Fix JavaExecableImp for Java11
