+++ src/main/java/com/diffplug/gradle/osgi/OsgiExecImp.java
@@ -17,10 +17,10 @@ package com.diffplug.gradle.osgi;
 
 
 import com.diffplug.gradle.FileMisc;
+import com.diffplug.gradle.JRE;
 import com.diffplug.gradle.JavaExecWinFriendly;
 import com.diffplug.gradle.SerializableMisc;
 import com.diffplug.gradle.ZipMisc;
-import com.diffplug.gradle.eclipserunner.JarFolderRunner;
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.InputStream;
@@ -43,7 +43,7 @@ class OsgiExecImp {
 			}
 		}
 		// look for our jar on the URLClassLoader path
-		for (URL url : JarFolderRunner.getClasspath(OsgiExecImp.class.getClassLoader())) {
+		for (URL url : JRE.getClasspath(OsgiExecImp.class.getClassLoader())) {
 			String name = url.getFile();
 			if (name != null) {
 				if (name.contains("/goomph")) {
